refactor: element.rs
