--- crates/uzumaki/src/elements/mod.rs
@@ -1,3 +1,0 @@
-pub mod input;
-pub mod text;
-pub mod view;
